requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-js-asset==0.1.1
django-mptt==0.9.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
factory-boy==2.9.2
Faker==0.8.8
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
six==1.11.0
text-unidecode==1.1
uritemplate==3.0.0
urllib3==1.22
